requirements.txt
awsiotsdk
python-dotenv~=1.1.1
numpy
smbus
libsvm-official
flask